refactor(e2e): Move helper functions to `libs/testing/e2e.py`
