+++ tasks/new_e2e_tests.py
@@ -13,7 +13,6 @@ import shutil
 import tempfile
 import threading
 from collections import defaultdict
-from collections.abc import Iterable
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from pathlib import Path
 
@@ -34,6 +33,7 @@ from tasks.libs.common.utils import (
     gitlab_section,
     running_in_ci,
 )
+from tasks.libs.testing.e2e import create_test_selection_regex, filter_only_leaf_tests
 from tasks.libs.testing.result_json import ActionType, ResultJson
 from tasks.test_core import DEFAULT_E2E_TEST_OUTPUT_JSON
 from tasks.testwasher import TestWasher
@@ -352,8 +352,8 @@ def run(
         "nocache": "-test.count=1" if not cache else "",
         "REPO_PATH": REPO_PATH,
         "commit": get_commit_sha(ctx, short=True),
-        "run": ('-test.run ' + _create_test_selection_regex(clean_run)) if run else "",
-        "skip": ('-test.skip ' + _create_test_selection_regex(clean_skip)) if skip else "",
+        "run": ('-test.run ' + create_test_selection_regex(clean_run)) if run else "",
+        "skip": ('-test.skip ' + create_test_selection_regex(clean_skip)) if skip else "",
         "test_run_arg": test_run_arg,
         "osversion": f"-osversion {osversion}" if osversion else "",
         "platform": f"-platform {platform}" if platform else "",
@@ -435,7 +435,7 @@ def run(
                     for package, _ in to_retry
                 }
                 e2e_module.test_targets = list(affected_packages)
-                args["run"] = '-test.run ' + _create_test_selection_regex([test for _, test in to_retry])
+                args["run"] = '-test.run ' + create_test_selection_regex([test for _, test in to_retry])
             else:
                 break
 
@@ -452,7 +452,7 @@ def run(
             os.path.relpath(package, "github.com/DataDog/datadog-agent/test/new-e2e/") for package, _ in to_teardown
         }
         e2e_module.test_targets = list(affected_packages)
-        args["run"] = '-test.run ' + _create_test_selection_regex([test for _, test in to_teardown])
+        args["run"] = '-test.run ' + create_test_selection_regex([test for _, test in to_teardown])
         env_vars["E2E_TEARDOWN_ONLY"] = "true"
         test_flavor(
             ctx,
@@ -690,60 +690,6 @@ def write_result_to_log_files(logs_per_test, log_folder, test_depth=1):
         sanitized_test_name = re.sub(r"[^\w_. -]", "_", test)
         with open(f"{log_folder}/{sanitized_package_name}.{sanitized_test_name}.log", "w") as f:
             f.write("".join(logs))
-
-
-def _create_test_selection_regex(test_names: list[str]) -> str:
-    """
-    Create a regex to exact-match the tests in the targets list.
-    Ex: ["TestFoo", "TestBar"] -> "^(?:TestFoo|TestBar)$"
-    """
-    if not test_names:
-        return ""
-
-    # Remove any whitespace and eventual ^$ surrounding the test names
-    processed_names = [name.strip().strip("^$") for name in test_names]
-
-    # Join them with a pipe to create an OR regex
-    regex_body = "|".join(processed_names)
-    if len(processed_names) > 1:
-        # If we have more than one test, we need to group them with a non-capturing group
-        regex_body = f"(?:{regex_body})"
-
-    return f'"^{regex_body}$"'
-
-
-def filter_only_leaf_tests(tests: Iterable[tuple[str, str]]) -> set[tuple[str, str]]:
-    """
-    Given some (package, test_name) tuples, return only the leaf tests.
-    A test is a leaf if it is not a parent of any other test in the list (within the same package).
-    """
-    # Sort tests by descending depth (number of '/' in test name)
-    tests_sorted = sorted(tests, key=lambda t: len(t[1].split('/')))
-    leaf_tests: set[tuple[str, str]] = set()
-    for candidate_test in tests_sorted:
-        # If none of the known leaf tests is a child of the candidate test, then the candidate test is a leaf
-        is_leaf = all(not is_child(candidate_test, known_leaf_test) for known_leaf_test in leaf_tests)
-        if is_leaf:
-            leaf_tests.add(candidate_test)
-    return leaf_tests
-
-
-def is_child(test1: tuple[str, str], test2: tuple[str, str]) -> bool:
-    """
-    Returns True if test1 is a child of test2.
-    Example: is_child(("pkg", "TestAgent/TestFeatureA"), ("pkg", "TestAgent")) == True
-    """
-    if test1[0] != test2[0]:
-        return False
-    splitted_test1 = test1[1].split('/')
-    splitted_test2 = test2[1].split('/')
-    if len(splitted_test2) > len(splitted_test1):
-        return False
-
-    for part1, part2 in zip(splitted_test1, splitted_test2, strict=False):
-        if part1 != part2:
-            return False
-    return True
 
 
 class TooManyLogsError(Exception):
